Introduce ResourceCaches object
