+++ components/resources/library/src/commonMain/kotlin/org/jetbrains/compose/resources/ResourceCaches.kt
@@ -1,20 +1,53 @@
 package org.jetbrains.compose.resources
 
-private val caches = mutableListOf<AsyncCache<*, *>>()
+import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 
-internal fun registerCache(cache: AsyncCache<*, *>) {
-    caches.add(cache)
+internal class AsyncCache<K, V> {
+    private val mutex = Mutex()
+    private val cache = mutableMapOf<K, Deferred<V>>()
+
+    init {
+        ResourceCaches.registerCache(this)
+    }
+
+    suspend fun getOrLoad(key: K, load: suspend () -> V): V = coroutineScope {
+        val deferred = mutex.withLock {
+            var cached = cache[key]
+            if (cached == null || cached.isCancelled) {
+                //LAZY - to free the mutex lock as fast as possible
+                cached = async(start = CoroutineStart.LAZY) { load() }
+                cache[key] = cached
+            }
+            cached
+        }
+        deferred.await()
+    }
+
+    fun clear() {
+        cache.clear()
+    }
 }
 
-/**
- * Clears any cached resources maintained internally by the system.
- *
- * It can be useful to release memory or reset cached resources that
- * may be changed or no longer be required.
- *
- * Note that frequent or unnecessary calls to this function may impact
- * performance by removing resources that might otherwise benefit from being cached.
- */
-fun dropResourceCaches() {
-   caches.forEach { it.clear() }
+object ResourceCaches {
+    private val caches = mutableListOf<AsyncCache<*, *>>()
+
+    internal fun registerCache(cache: AsyncCache<*, *>) = caches.add(cache)
+
+    /**
+     * Clears any cached resources maintained internally by the system.
+     *
+     * It can be useful to release memory or reset cached resources that
+     * may be changed or no longer be required.
+     *
+     * Note that frequent or unnecessary calls to this function may impact
+     * performance by removing resources that might otherwise benefit from being cached.
+     */
+    fun drop() {
+        caches.forEach { it.clear() }
+    }
 }
